Stop yield benchmark from timing out
seL4_SchedContext_Yield is not quite the same as seL4_Yield which it
replaces as it sacrifices the rest of the current timeslice and budget
is not available again until the period has passed.

This chane fixes the yield benchmark for the RT branch by reducing the
timeslice of the yielding threads so that the benchmark completes
quickly.