sel4bench: disable fault benchmark for armv6
The fault handler benchmark works by
triggering an undefined instruction. On armv6,
undefined instruction is leveraged by the benchmarks
(along with CONFIG_DANGEROUS_CODE_INJECTION) to
read the cycle counter quickly, as armv6 does
not allow the cycle counter to be read from user level.

This means triggering an undefined instruction breaks the fault benchmark on
armv6, so the solution for now is to disable it.