Cache timer frame caps
The simple implementation, due to changes in the kernel API, is not
able to allocate/return the same device frame cap multiple times.
This change caches the frame cap for the timer allowing us to give
it to multiple benchmark instances.

Additionally adds a check for failure to copy the frame cap