sel4bench: HPET for x86, untyped for all timers
This commit updates sel4bench for changes to seL4_libs
where the HPET is used as the default timer for x86 and timer
frames are now allocated from the allocator, not simple.

Instead of passing a cap to a frame to sel4bench child apps,
we pass the untyped with the paddr for the timer to give to the
child app allocator.