Increase untyped capabilities
Current configs do not provide for enough untyped capabilities for some platforms.
With devices now being passed as untypeds (and not separate frame caps)
there are not enough untyped slots to fit call the capabilities. This
is causing the untyped capabilities for some devices to be dropped,
causing tests to fail