Clobber message registers in fault receive macro
The DO_REPLY_RECV_1 macro replies with a single message word and then
waits on an endpoint for a message. This message received, however, is
a fault message and so contains message data. The first 4 words of this
will get placed in the message registers. Currently this stub retrieves
the first word and ignores the others. This is incorrect as the other
registers will be being overwritten by this message.

This commit adds the other 3 message registers as clobbers to capture
the fact that the kernel will overwrite these.