sel4bench: use slab allocator for critical objects
This change alters sel4bench applications to declare objects
that the kernel will touch during the benchmark to be declared
with the slab allocator such that they can be colocated to
increase locality and decrease the chance of collisions.

Since the slab allocator does not yet support free, it also
rearranges some benchmarks to avoid constantly freeing and allocating
new threads and processes. This is desirable as the allocations were
not neccessary.