SELFOUR-1048: Fix page mapping benchmark crashing when running the 8th test
Prior to this commit the 8th test would fail as it would have enough pages such that
the (incorrect) formula for calculating the number of page tables of report that another
page table was needed. This would result in this page table, after masking out irrelevant
parts of the address, having the same location as the previous page table, thus leading
to an error and the whole benchmark crashes.

This commit corrects the `PAGE_PER_TABLE` macro and allows all the tests to continue