Use seL4_Wait instead of seL4_Recv where possible
On the master kernel seL4_Wait just calls seL4_Recv and does not return
a message tag. On the mcs kernel, seL4_Wait is a first class system call which
waits on an object without requiring a reply object, but does return a
message tag. Therefore seL4_Wait can be
used equivalently on both branches as long as the tag is not required.

This change replaces calls to seL4_Recv where the return value is
ignored with seL4_Wait.