Only init timers for benchmarks that need them
Prevously each benchmark would init the timers when benchmark_get_env
was called.

This change puts the timer init into a separate function which
benchmarks must call if they want to use a timer, and updates the
benchmarks that need to a timer to call it.

This prevents unnecessary noise in the benchmarks from unused timers.