Spawn process without runtime but with TLS
When sel4bench spawns new processes it jumps directly to a function
without allowing a runtime to initialise the process. Doing this leads
to a the running thread to start without an initialised TLS region.

The changes to seL4 regarding TLS require a TLS region to be initialised
if the IPC buffer is going to be accessed from user-space.

To allow this technique to continue to be used a custom process
launcher has been added that initialises the TLS region for the newly
created process within its stack allocation.