CMake: Avoid settings that are not defined yet
In order to remove the requirement for invoking cmake multiple times to
configure a build, we need to only test settings that have been
previously set. While the settings.cmake file is being processed, many
kernel boolean properties are not set yet, and we instead need to test
the string variants.