Force FPU state to only be restored when used
This ensures that during benchmarks where the FPU state isn't being used
that the FPU state does not get restored. This ensures that context
switches in benchmarks accurately show the case where FPU restore is not
performed.

It may be resonable to ensure that there are benchmarks showing the
overhead where the FPU is forced to be restored.