libsel4benchsupport: fix child process stack init
benchmark_spawn_process would move the "stack top" pointer of a process
down every time it was called. This is OK for most use cases, where a
process is used once and then discarded, but for the page_mapping
benchmark the same process is used repeatedly, calling
sel4bench_spawn_process every time it's started.

This causes the available stack size to grow smaller and smaller,
resulting in the page_mapping benchmark eventually overrunning the
stack and crashing.

Fix this by ensuring benchmark_spawn_process leaves the stack top
intact.