smp: Fix checkpointing
Previously, threads were being checkpointed before they even start.
This was a problem for certain architectures (e.g. x86_64) as the
checkpoint function required threads to be blocked inside the kernel.
Those architectures had special kernel ABI behaviours and hence the
checkpointing at the start would break those behaviours and the
benchmark would misbehave.

This commit fixes the checkpointing so that checkpoints are taken when
threads are blocked on 'seL4_Wait'.