add project options to switch benchmarks on-off
easy-settings.cmake was amended by configuration options to choose
benchmark applications to be included in the image.
The settings should be the same as local defaults in corresponding
app directories.

easy-settings.cmake introduces simplified names of the options.
Matching of the simplified names to the option names in local
directories can be found in settings.cmake
(Ex. "SCHED" - "AppSchedulerBench")

Signed-off-by: Nataliya Korovkina <malus.brandywine@gmail.com>