SCHED: Don't let spawned processes crash
The function will return and crash because there is nothing to
return to and apparantly the sel4runtime doesn't implement a
soft landing.

Test still works with crashing processes and with debugging
disabled it's invisible. But very confusing when trying to
debug any kind of other problems and enabling debugging.

Signed-off-by: Indan Zupancic <indan@nul.nu>