Eliminate gcc-14 warning on const type mismatch
gcc-14 complains about the const difference in the type of argv between
the various main() functions (char **argv) and benchmark_get_env() which
has char const *const *argv.

Reduce benchmark_get_env to char **argv to eliminate the warning.
This is less information, but const pointers are weak in C anyway, so
there is not much point updating all the types of all main() functions
instead.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>